feat: add support for specifying a single collection in Mongo struct and update GetMongoClient function to handle dynamic collection name
The Mongo struct now includes a new field called "Collection" which allows specifying a single collection to use. The GetMongoClient function has been updated to handle this dynamic collection name by retrieving the collection from the Mongo struct before returning the client. This allows for more flexibility in working with different collections in the MongoDB database. Additionally, the deprecated comment has been added to indicate that the RealMongoOperations.GetMongoClient function is deprecated as of ConfigBuilder v0.5.0.