feat(mongo): Upgrade to MongoDB v2.2.1 driver
Upgrade the MongoDB driver to the latest version (v2.2.1) to take advantage of
the latest features and bug fixes. This includes changes to the import paths
and the way the client is created. The `GetMongoClient` function has been
updated to use the new driver version and set the read preference to
`SecondaryPreferred` to improve read availability in a replica set
environment.